+++ CHANGES.rst
@@ -5,7 +5,8 @@ Changelog
 1.0a2 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Upgrade to be compatible with eea.facetednavigation 10 and above
+  [mpeeters]
 
 
 1.0a1 (2018-03-01)